Update authorization through JWT
Create outside `ResourceServer` application to creating tokens (to authentication) by login and password
