--- src/main/java/com/CezaryZal/config/Beans.java
@@ -1,27 +1,0 @@
-package com.CezaryZal.config;
-
-import org.springframework.boot.web.servlet.FilterRegistrationBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-
-import java.util.Collections;
-
-@Service
-public class Beans {
-
-    @Bean
-    public PasswordEncoder passwordEncoder(){
-        return new BCryptPasswordEncoder();
-    }
-
-//    @Bean
-//    public FilterRegistrationBean filterRegistrationBean(){
-//        FilterRegistrationBean filterRegistration = new FilterRegistrationBean();
-//        filterRegistration.setFilter(new JwtFilter());
-//        filterRegistration.setUrlPatterns(Collections.singleton("/api/*"));
-//
-//        return filterRegistration;
-//    }
-}
