+++ src/main/java/com/CezaryZal/security/JwtFilter.java
@@ -1,35 +1,69 @@
-//package com.CezaryZal.security;
+package com.CezaryZal.security;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jws;
+import io.jsonwebtoken.Jwts;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
+
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.util.Collections;
+import java.util.Set;
+
+public class JwtFilter extends BasicAuthenticationFilter {
+
+    public JwtFilter(AuthenticationManager authenticationManager) {
+        super(authenticationManager);
+    }
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain chain) throws IOException, ServletException {
+
+        String header = request.getHeader("Authorization");
+        UsernamePasswordAuthenticationToken authResult = getAuthenticationByToken (header);
+        SecurityContextHolder.getContext().setAuthentication(authResult);
+        //po pomyślnej autoryzacji filter przekazuje rządanie dalej
+        chain.doFilter(request, response);
+    }
+
+    //Szyfrownianie symetryczne
+    private UsernamePasswordAuthenticationToken getAuthenticationByToken(String header) {
+        Jws<Claims> claimsJws = Jwts
+                .parser()
+                .setSigningKey("dFKYX}&[K;'{{;q".getBytes())
+                .parseClaimsJws(header.replace("Bearer ", ""));
+
+        String userName = claimsJws.getBody().get("name").toString();
+        String role = claimsJws.getBody().get("role").toString();
+
+        Set<SimpleGrantedAuthority> simpleGrantedAuthorities = Collections.singleton(new SimpleGrantedAuthority(role));
+
+        UsernamePasswordAuthenticationToken usernamePasswordAuthenticationToken =
+                new UsernamePasswordAuthenticationToken(userName, null, simpleGrantedAuthorities);
+        return usernamePasswordAuthenticationToken;
+    }
+
+//        public AuthenticationResponse authenticate(AuthenticationRequest authenticationRequest) {
+//        UsernamePasswordAuthenticationToken authenticationToken =
+//                new UsernamePasswordAuthenticationToken(authenticationRequest.getUsername(), authenticationRequest.getPassword());
 //
-//import io.jsonwebtoken.Claims;
-//import io.jsonwebtoken.Jwts;
+//        Authentication authentication = authenticationManager.authenticate(authenticationToken);
 //
-//import javax.servlet.FilterChain;
-//import javax.servlet.ServletException;
-//import javax.servlet.ServletRequest;
-//import javax.servlet.ServletResponse;
-//import javax.servlet.http.HttpServletRequest;
-//import java.io.IOException;
+//        if (authentication.isAuthenticated()) {
+//            String jwt = jwtBuilder.build(authentication, AuthenticationConstants.JWT_EXPIRE_IN_MINUTES);
 //
-//public class JwtFilter implements javax.servlet.Filter {
+//            RefreshToken refreshToken = refreshTokenService.buildRefreshToken(authentication, AuthenticationConstants.REFRESH_TOKEN_EXPIRE_IN_MINUTES);
+//            refreshTokenService.addToken(refreshToken);
 //
-//    @Override
-//    public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
-//        HttpServletRequest httpServletRequest = (HttpServletRequest) servletRequest;
-//        String header = httpServletRequest.getHeader("authorization");
-//
-//        if(httpServletRequest == null || !header.startsWith("Bearer ")){
-//            throw new ServletException("Wrong or empty header");
-//        } else {
-//            try {
-//                String token = header.substring(7);
-//                Claims claims = Jwts.parser().setSigningKey("admin123").parseClaimsJws(token).getBody();
-//                servletRequest.setAttribute("claims", claims);
-//            } catch (Exception e){
-//                throw new ServletException("Wrong key");
-//            }
-//
+//            return new AuthenticationResponse(jwt, refreshToken.getValue());
 //        }
-//        filterChain.doFilter(servletRequest, servletResponse);
-//
+//        throw new LoginException("Invalid username or password");
 //    }
-//}
+}
