--- src/main/java/com/CezaryZal/security/UserPrincipal.java
@@ -1,67 +1,0 @@
-//package com.CezaryZal.security;
-//
-//import com.CezaryZal.user.User;
-//import org.springframework.security.core.GrantedAuthority;
-//import org.springframework.security.core.authority.SimpleGrantedAuthority;
-//import org.springframework.security.core.userdetails.UserDetails;
-//
-//import java.util.ArrayList;
-//import java.util.Collection;
-//import java.util.List;
-//
-//public class UserPrincipal implements UserDetails {
-//
-//    private User user;
-//
-//    public UserPrincipal(User user) {
-//        this.user = user;
-//    }
-//
-//    @Override
-//    public Collection<? extends GrantedAuthority> getAuthorities() {
-//        List<GrantedAuthority> authorities = new ArrayList<>();
-//
-//        //Extract list of permissions (name)
-//        this.user.getPermissionList().forEach(p ->{
-//            GrantedAuthority authority = new SimpleGrantedAuthority(p);
-//            authorities.add(authority);
-//        });
-//
-//        //Extract list of roles (ROLE_name)
-//        this.user.getRoleList().forEach(r ->{
-//            GrantedAuthority authority = new SimpleGrantedAuthority("ROLE_" + r);
-//            authorities.add(authority);
-//        });
-//        return authorities;
-//    }
-//
-//    @Override
-//    public String getPassword() {
-//        return user.getPassword();
-//    }
-//
-//    @Override
-//    public String getUsername() {
-//        return user.getLoginName();
-//    }
-//
-//    @Override
-//    public boolean isAccountNonExpired() {
-//        return true;
-//    }
-//
-//    @Override
-//    public boolean isAccountNonLocked() {
-//        return true;
-//    }
-//
-//    @Override
-//    public boolean isCredentialsNonExpired() {
-//        return true;
-//    }
-//
-//    @Override
-//    public boolean isEnabled() {
-//        return this.user.isActive();
-//    }
-//}
