--- src/main/java/com/CezaryZal/user/UserToken.java
@@ -1,23 +1,0 @@
-package com.CezaryZal.user;
-
-public class UserToken {
-
-    private String login;
-    private String password;
-
-    public String getLogin() {
-        return login;
-    }
-
-    public void setLogin(String login) {
-        this.login = login;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
